move degeneric derive into separate module
