+++ src/degeneric/mod.rs
@@ -1,0 +1,7 @@
+mod attribute;
+mod degeneric;
+mod field;
+mod generics;
+mod type_tools;
+
+pub use self::degeneric::process_struct;
